+++ go-client-sdk/pkg/models/operations/usageexamplepost.go
@@ -201,6 +201,12 @@ type UsageExamplePostRequest struct {
 	RequestBody *UsageExamplePostRequestBody `request:"mediaType=application/json"`
 	// An bigint parameter
 	BigintParameter *big.Int `queryParam:"style=form,explode=true,name=bigintParameter"`
+	// An bigint parameter
+	BigintParameterOptional *big.Int `queryParam:"style=form,explode=true,name=bigintParameterOptional"`
+	// An bigint parameter
+	BigintStrParameter types.BigInt `queryParam:"style=form,explode=true,name=bigintStrParameter"`
+	// An bigint parameter
+	BigintStrParameterOptional *types.BigInt `queryParam:"style=form,explode=true,name=bigintStrParameterOptional"`
 	// A boolean parameter
 	BoolParameter bool `queryParam:"style=form,explode=true,name=boolParameter"`
 	// A date parameter
@@ -209,6 +215,12 @@ type UsageExamplePostRequest struct {
 	DateTimeParameter time.Time `queryParam:"style=form,explode=true,name=dateTimeParameter"`
 	// A decimal parameter
 	DecimalParameter float64 `queryParam:"style=form,explode=true,name=decimalParameter"`
+	// A decimal parameter
+	DecimalParameterOptional *float64 `queryParam:"style=form,explode=true,name=decimalParameterOptional"`
+	// A decimal parameter
+	DecimalStrParameter string `queryParam:"style=form,explode=true,name=decimalStrParameter"`
+	// A decimal parameter
+	DecimalStrParameterOptional *string `queryParam:"style=form,explode=true,name=decimalStrParameterOptional"`
 	// A double parameter
 	DoubleParameter float64 `queryParam:"style=form,explode=true,name=doubleParameter"`
 	// An enum parameter
@@ -243,6 +255,27 @@ func (o *UsageExamplePostRequest) GetBigintParameter() *big.Int {
 	return o.BigintParameter
 }
 
+func (o *UsageExamplePostRequest) GetBigintParameterOptional() *big.Int {
+	if o == nil {
+		return nil
+	}
+	return o.BigintParameterOptional
+}
+
+func (o *UsageExamplePostRequest) GetBigintStrParameter() types.BigInt {
+	if o == nil {
+		return big.NewInt(0)
+	}
+	return o.BigintStrParameter
+}
+
+func (o *UsageExamplePostRequest) GetBigintStrParameterOptional() *types.BigInt {
+	if o == nil {
+		return nil
+	}
+	return o.BigintStrParameterOptional
+}
+
 func (o *UsageExamplePostRequest) GetBoolParameter() bool {
 	if o == nil {
 		return false
@@ -269,6 +302,27 @@ func (o *UsageExamplePostRequest) GetDecimalParameter() float64 {
 		return 0.0
 	}
 	return o.DecimalParameter
+}
+
+func (o *UsageExamplePostRequest) GetDecimalParameterOptional() *float64 {
+	if o == nil {
+		return nil
+	}
+	return o.DecimalParameterOptional
+}
+
+func (o *UsageExamplePostRequest) GetDecimalStrParameter() string {
+	if o == nil {
+		return ""
+	}
+	return o.DecimalStrParameter
+}
+
+func (o *UsageExamplePostRequest) GetDecimalStrParameterOptional() *string {
+	if o == nil {
+		return nil
+	}
+	return o.DecimalStrParameterOptional
 }
 
 func (o *UsageExamplePostRequest) GetDoubleParameter() float64 {