+++ go-client-sdk/sdk.go
@@ -277,7 +277,7 @@ func New(opts ...SDKOption) *SDK {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "0.1.0",
-			SDKVersion:        "1.31.0",
+			SDKVersion:        "1.31.1",
 			GenVersion:        "2.101.0",
 			Globals: map[string]map[string]map[string]interface{}{
 				"parameters": {},