+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDKConfiguration.java
@@ -35,8 +35,8 @@ class SDKConfiguration {
 	}};
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
-	public String sdkVersion = "1.31.5";
-	public String genVersion = "2.96.9";
+	public String sdkVersion = "1.31.6";
+	public String genVersion = "2.101.0";
 	public Map<String, Map<String, Map<String, Object>>> globals = new HashMap<String, Map<String, Map<String, Object>>>(){{
 		put("parameters", new HashMap<String, Map<String, Object>>());
 	}};