+++ php-client-sdk/src/RequestBodies.php
@@ -1350,6 +1350,47 @@ class RequestBodies
     }
 	
     /**
+     * requestBodyPostComplexNumberTypes
+     * 
+     * @param \OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesRequest $request
+     * @return \OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesResponse
+     */
+	public function requestBodyPostComplexNumberTypes(
+        ?\OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesRequest $request,
+    ): \OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesResponse
+    {
+        $baseUrl = Utils\Utils::templateUrl($this->sdkConfiguration->getServerUrl(), $this->sdkConfiguration->getServerDefaults());
+        $url = Utils\Utils::generateUrl($baseUrl, '/anything/requestBodies/post/{pathBigInt}/{pathBigIntStr}/{pathDecimal}/{pathDecimalStr}/complex-number-types', \OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesRequest::class, $request, $this->sdkConfiguration->globals);
+        
+        $options = ['http_errors' => false];
+        $body = Utils\Utils::serializeRequestBody($request, "complexNumberTypes", "json");
+        if ($body !== null) {
+            $options = array_merge_recursive($options, $body);
+        }
+        $options = array_merge_recursive($options, Utils\Utils::getQueryParams(\OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesRequest::class, $request, $this->sdkConfiguration->globals));
+        $options['headers']['Accept'] = 'application/json';
+        $options['headers']['x-speakeasy-user-agent'] = sprintf('speakeasy-sdk/%s %s %s %s', $this->sdkConfiguration->language, $this->sdkConfiguration->sdkVersion, $this->sdkConfiguration->genVersion, $this->sdkConfiguration->openapiDocVersion);
+        
+        $httpResponse = $this->sdkConfiguration->securityClient->request('POST', $url, $options);
+        
+        $contentType = $httpResponse->getHeader('Content-Type')[0] ?? '';
+
+        $response = new \OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypesResponse();
+        $response->statusCode = $httpResponse->getStatusCode();
+        $response->contentType = $contentType;
+        $response->rawResponse = $httpResponse;
+        
+        if ($httpResponse->getStatusCode() === 200) {
+            if (Utils\Utils::matchContentType($contentType, 'application/json')) {
+                $serializer = Utils\JSON::createSerializer();
+                $response->requestBodyPostComplexNumberTypes200ApplicationJSONObject = $serializer->deserialize((string)$httpResponse->getBody(), 'OpenAPI\OpenAPI\Models\Operations\RequestBodyPostComplexNumberTypes200ApplicationJSON', 'json');
+            }
+        }
+
+        return $response;
+    }
+	
+    /**
      * requestBodyPostEmptyObject
      * 
      * @param \OpenAPI\OpenAPI\Models\Operations\RequestBodyPostEmptyObjectRequestBody $request