+++ php-client-sdk/src/SDKConfiguration.php
@@ -36,8 +36,8 @@ class SDKConfiguration
 	];
 	public string $language = 'php';
 	public string $openapiDocVersion = '0.1.0';
-	public string $sdkVersion = '1.31.5';
-	public string $genVersion = '2.96.9';
+	public string $sdkVersion = '1.31.6';
+	public string $genVersion = '2.101.0';
 	/** @var array<string, array<string, array<string, mixed>>> */
 	public ?array $globals = [
     	'parameters' => []