+++ python-client-sdk/src/sdk/models/errors/error.py
@@ -14,6 +14,6 @@ class Error(Exception):
     r"""Internal Server Error"""
     code: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('code'), 'exclude': lambda f: f is None }})
     message: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('message'), 'exclude': lambda f: f is None }})
-    
+
     def __str__(self) -> str:
         return self.message