+++ python-client-sdk/src/sdk/models/errors/statusgetxspeakeasyerrors.py
@@ -16,6 +16,6 @@ class StatusGetXSpeakeasyErrors501ApplicationJSON(Exception):
     code: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('code'), 'exclude': lambda f: f is None }})
     message: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('message'), 'exclude': lambda f: f is None }})
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'exclude': lambda f: True }})
-    
+
     def __str__(self) -> str:
         return utils.marshal_json(self)